+++ engine/include/jc2d.h
@@ -1,0 +1,1 @@
+#include "../src/core/application.h"
