+++ engine/src/core/logger/logger.cpp
@@ -1,5 +1,5 @@
-#include <engine/logger.h>
-#include <engine/asserts.h>
+#include "logger.h"
+#include "../../asserts.h"
 
 // TODO: temporary
 #include <stdio.h>
